价格属性的类型调整,全部由int改成double

+++ src/main/java/edu/graduate/messhall/bean/TblCart.java
@@ -41,7 +41,7 @@ public class TblCart{
     @Transient
     static private boolean checked = false;
     @Transient
-    private int foodPrice;
+    private double foodPrice;
     @Transient
     private String foodName;
